umem/profile.rs: extract OwnerProfile telemetry counters (#986 Phase 1)
Second slice of the umem.rs decomposition (after #1019 Phase 0
extracted MmapArea). Pulls the per-binding telemetry counters
out: TX/RX latency histograms, drain invocation counts,
redirect-acquire histograms, owner/peer pps.

Mechanical changes:
- New umem/profile.rs (~190 LOC): OwnerProfileOwnerWrites
  (#[repr(align(64))]) and OwnerProfilePeerWrites (#[repr(align(64))])
  structs + their inline new() constructors. Single-writer
  (owner worker) for OwnerWrites, peer-writer (cross-binding
  redirect completion path) for PeerWrites; all stats use
  Ordering::Relaxed.
- umem/mod.rs declares mod profile and pub(in crate::afxdp)
  re-exports both struct types so existing call sites continue
  to reach them via crate::afxdp::umem::OwnerProfile*.
- Visibility widened pub(super) → pub(in crate::afxdp) on
  struct + fields + impl new() to preserve cross-module reach.

After this:
- umem/mod.rs: 2,344 LOC (down from 2,525 after Phase 0; Phase 2
  candidates: BindingLiveState extraction, ring management).
- umem/profile.rs: 192 LOC.

cargo test --bins 865/0/2.
